requirements.txt
starlette
uvicorn

django
django-extensions
django-rest-framework

djangorestframework-simplejwt

names
python-keycloak